pyproject.toml
[tool.poetry]
name = "sample_blog"
version = "0.1.0"
description = "Demo App Project built using Django4 by Example by Antonio Mele"
authors = ["David J Nevin <content@davidjnevin.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.6"
dj-database-url = "^1.0.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
Django = "^4.1.1"
django-taggit = "^3.0.0"
Markdown = "3.4.1"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
